\name{barplot3d-package}
\alias{barplot3d-package}
\alias{barplot3d}
\docType{package}
\title{
\packageTitle{barplot3d}
}
\description{
\packageDescription{barplot3d}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{barplot3d}
\packageIndices{barplot3d}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{barplot3d}

Maintainer: \packageMaintainer{barplot3d}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
