% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/barplot3d.R
\name{barplot3d}
\alias{barplot3d}
\title{Barplot 3D}
\usage{
barplot3d(z, dimensions, scalexy = 1, scalez = 1, gap = 0.2, alpha = 1,
  colors = NA, linecol = "black")
}
\arguments{
\item{z}{Numeric vector of the heights of each column}

\item{dimensions}{Numeric vector of length two; the number of rows and columns of the base of the plot}

\item{scalexy}{Numeric value used to scale the x and y coordinates (i.e. the "fatness" of the columns). Also scales the gap between columns}

\item{scalez}{Numeric value used to scale the z input vector. Useful for playing with data, but not recommended}

\item{gap}{Numeric value determining the gap between columns; this gap is identical in the x and y dimensions}

\item{alpha}{Numeric value determining the transparency of all column sides. Ranges beween 0 (invisible) and 1 (opaque)}

\item{colors}{Colour of columns.  Either a single colour or a character vector of colours, one for each column}

\item{linecol}{Colour of edges of columns. A single colour.}
}
\value{
Draws 3D barplot; only the most recent object IDs will be returned.
}
\description{
Creates 3D barplots, with many configurable options. For detailed usage, please read the accompanying vignette.
}
\details{
Produces a 3D barplot using a numeric vector of values. The \code{\link{rgl}} package is used to draw the plots.
Further details are available in the accompanying vignette.
}
\examples{
## The vignette provides far more in-depth explanation and examples ##
blah blah blah
}
\author{
Christopher Wardell \email{r@cpwardell.com}
}

